+++ src/mode.rs
@@ -1,3 +1,4 @@
+pub mod command;
 pub mod editor;
 pub mod explorer;
 pub mod picker;
@@ -14,6 +15,7 @@ pub enum Mode {
     QueryNormal,
     QueryInsert,
     Results,
+    Command,
 }
 
 impl Mode {
@@ -24,6 +26,7 @@ impl Mode {
             Mode::QueryInsert => editor::insert::handle_key(key, app),
             Mode::Explorer => explorer::handle_key(key, app),
             Mode::Results => results::handle_key(key, app),
+            Mode::Command => command::handle_key(key, app),
         }
     }
 
@@ -34,6 +37,7 @@ impl Mode {
             Mode::QueryNormal => "NORMAL",
             Mode::QueryInsert => "INSERT",
             Mode::Results => "RESULTS",
+            Mode::Command => "COMMAND",
         }
     }
 }